requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.7
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
